adds more tests for attack point optimizers
